Almost done with Infdev 20100327
- Chunks are now a dedicated Class
- Chunks now hold and generate a heightmap
- Chunks now have a getBlock function
- Inherited the chunk population logic from Infdev 20100327. Unsure if this is better, but its much more readable
- PrepareGenerator is now directly part of the Generator Init. It used to be different back when a unique generator was used for the initial spawn chunks
-  Made a lot of stuff unique pointers to avoid some dangling pointer messes
- Renamed the infdev noise functions to be in line with their slightly newer equivalents
- Fixed a MAJOR oversight when loading the seed from server.properties, always interpreting the read number as an int (32-Bit) instead of a long (64-Bit), resulting in a truncated and just wrong generation
- Infdev20100327 Terrain Generation matches 1:1
- Infdev20100327 Ore Generation matches 1:1
- Infdev20100327 Tree generation is a WIP
- The farlands do not generate

+++ src/server.cpp
@@ -152,7 +152,7 @@ void Server::LoadConfig() {
 	} else {
 		GlobalConfig::Instance().LoadFromDisk();
 		chunkDistance = GlobalConfig::Instance().GetAsNumber<int>("view-distance");
-		seed = GlobalConfig::Instance().GetAsNumber<int>("level-seed");
+		seed = GlobalConfig::Instance().GetAsNumber<int64_t>("level-seed");
 		maximumPlayers = GlobalConfig::Instance().GetAsNumber<int>("max-players");
 		whitelistEnabled = GlobalConfig::Instance().GetAsBoolean("white-list");
 	}
